pyproject.toml
[tool.poetry]
name = "NetInsight"
version = "0.1.0"
description = "NetDevOPS Web Framework based on FastAPI"
authors = ["wangxin.jeffry"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.77.1"
uvicorn = "^0.17.6"
SQLAlchemy = "^1.4.36"
alembic = "^1.7.7"
Faker = "^13.11.1"
sentry-sdk = "^1.5.12"
toml = "^0.10.2"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy-Utils = "^0.38.2"
pydantic = {extras = ["email"], version = "^1.9.0"}
python-dotenv = "^0.20.0"
unicode = "^2.8"
python-multipart = "^0.0.5"
autoflake = "^1.4"
black = "^22.3.0"
isort = "^5.10.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
